requirement.txt
absl-py==1.0.0
aggdraw==1.3.15
alabaster==0.7.12
anycache==2.0.7
anyio==3.4.0
appdirs==1.4.4
aquvitae==0.1.7
argon2-cffi==21.1.0
arrow==1.2.1
astroid==2.9.3
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.2.0
audioread==2.1.9
autopep8==1.6.0
babel==2.9.1
backcall==0.2.0
binaryornot==0.4.4
black==21.12b0
bleach==4.1.0
cachetools==4.2.4
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
cloudpickle==2.0.0
cookiecutter==1.7.3
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
dm-tree==0.1.6
docutils==0.17.1
entrypoints==0.3
filelock==3.8.0
fire==0.4.0
flake8==4.0.1
flatbuffers==1.12
fonttools==4.28.2
future==0.18.2
fxpmath==0.4.5
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
imageio==2.24.0
imagesize==1.3.0
importlib-metadata==4.8.2
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.7.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
jedi==0.18.1
jeepney==0.7.1
jellyfish==0.9.0
jinja2==3.0.3
jinja2-time==0.2.0
joblib==1.1.0
json5==0.9.6
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.12.0
jupyterlab==3.2.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
kapre==0.3.7
keras==2.7.0
keras-bert==0.87.0
keras-embed-sim==0.9.0
keras-layer-normalization==0.15.0
keras-multi-head==0.28.0
keras-pos-embd==0.12.0
keras-position-wise-feed-forward==0.7.0
keras-preprocessing==1.1.2
keras-self-attention==0.50.0
keras-transformer==0.39.0
keras-tuner==1.1.2
keras2onnx==1.7.0
keyring==23.5.0
kiwisolver==1.3.2
kt-legacy==1.0.4
lazy-object-proxy==1.7.1
libclang==12.0.0
librosa==0.9.1
llvmlite==0.38.0
markdown==3.3.6
markupsafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
netron
networkx==2.6.3
notebook==6.4.6
np-utils==0.5.12.1
numba==0.55.1
numpy==1.21.4
numpydoc==1.1.0
nvidia-pyindex==1.0.9
oauthlib==3.1.1
onnx==1.10.2
onnx-graphsurgeon==0.3.19
onnx-simplifier==0.3.6
onnx-tf==1.9.0
onnx2keras==0.0.24
onnxconverter-common==1.9.0
onnxoptimizer==0.2.6
onnxruntime==1.10.0
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pafy==0.5.5
pandas==1.3.4
pandocfilters==1.5.0
parse==1.6.5
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.4.0
pip==22.3.1
platformdirs==2.4.1
pluggy==1.0.0
pooch==1.6.0
poyo==0.5.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
pygments==2.10.0
pylint==2.12.2
pyls-spyder==0.4.0
pyparser==1.0
pyparsing==3.0.6
pyqt5==5.12.3
pyqt5-sip==12.9.0
pyqtwebengine==5.12.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-graphviz==0.20.1
python-lsp-black==1.0.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.3.3
python-slugify==5.0.2
pytz==2021.3
pywavelets==1.4.1
pyxdg==0.27
pyyaml==6.0
pyzmq==22.3.0
qdarkstyle==3.0.2
qkeras==0.9.0
qstylizer==0.2.1
qtawesome==1.1.1
qtconsole==5.2.2
qtpy==2.0.0
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rope==0.22.0
rsa==4.8
rtree==0.9.7
scikit-image==0.19.3
scikit-learn==1.0.1
scipy==1.7.3
secretstorage==3.3.1
send2trash==1.8.0
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.10.3.post1
sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.2.1
spyder-kernels==2.2.1
tdqm==0.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-addons==0.16.1
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-model-optimization==0.7.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
textdistance==4.2.2
tf2onnx==1.9.3
threadpoolctl==3.0.0
three-merge==0.1.1
tifffile==2022.10.10
tinycss2==1.1.1
toml==0.10.2
tomli==1.2.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typeguard==2.13.3
typing-extensions==4.0.0
ujson==5.1.0
urllib3==1.26.7
visualkeras==0.0.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
wurlitzer==3.0.2
yapf==0.32.0
youtube-dl==2021.6.6
zipp==3.6.0